Removed the LoginView class and moved the UI setup code to LoginViewController.
